# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FinReport Agent M1 API
  version: 0.1.0
  description: |
    M1.01–M1.12 的认证、上传、任务和 SSE 接口。错误响应遵循 RFC 9457。
servers:
  - url: http://localhost:8080
paths:
  /api/v1/reports/upload:
    post:
      summary: 上传 PDF 财报并创建或复用处理任务
      description: |
        成功始终保持 HTTP 201。相同用户与同一 PDF MD5 在任务为排队、运行或成功时会返回已有 task/report；
        FAILED 或 CANCELLED 的历史任务仅可用新的 Idempotency-Key 创建重试任务。Idempotency-Key 按用户隔离。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, companyCode, companyName, reportType, reportPeriod]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF，最大 50 MB。
                companyCode:
                  type: string
                  example: '600519'
                companyName:
                  type: string
                  example: 贵州茅台
                reportType:
                  type: string
                  example: ANNUAL
                reportPeriod:
                  type: string
                  example: '2024-12-31'
      responses:
        '201':
          description: 新建任务或符合去重规则的既有任务。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '409':
          description: Idempotency-Key 冲突。
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                keyReused:
                  value:
                    type: https://finreport.example/errors/IDEMPOTENCY_KEY_REUSED
                    title: Conflict
                    status: 409
                    detail: 同一个 Idempotency-Key 不能用于不同的文件
                inProgress:
                  value:
                    type: https://finreport.example/errors/IDEMPOTENCY_REQUEST_IN_PROGRESS
                    title: Conflict
                    status: 409
                    detail: 同一 Idempotency-Key 的上传请求仍在处理中
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
...
  schemas:
    UploadResponse:
      type: object
      required: [taskId, reportId, status]
      properties:
        taskId:
          type: string
        reportId:
          type: integer
          format: int64
        status:
          type: string
          example: QUEUED
...
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        traceId:
          type: string
...